+++ huggingface/third-party-models/video-to-video/cosmos_transfer1/rbln_auxiliary/sam2/sam2_model.py
@@ -131,7 +131,7 @@ class VideoSegmentationModel:
         results = self.processor.post_process_grounded_object_detection(
             outputs,
             inputs.input_ids,
-            box_threshold=0.15,
+            threshold=0.15,
             text_threshold=0.25,
             target_sizes=[image.size[::-1]],
         )
@@ -146,7 +146,7 @@ class VideoSegmentationModel:
             results = self.processor.post_process_grounded_object_detection(
                 outputs,
                 inputs.input_ids,
-                box_threshold=0.1,
+                threshold=0.1,
                 text_threshold=0.1,
                 target_sizes=[image.size[::-1]],
             )