+++ huggingface/transformers/visual-document-retrieval/colpali/compile.py
@@ -48,8 +48,18 @@ def main():
     rbln_model = RBLNColPaliForRetrieval.from_model(
         model,
         export=True,  # export a PyTorch model to RBLN model with optimum
-        rbln_max_seq_lens=8192,  # default "max_position_embeddings"
+        rbln_batch_size=2,
         rbln_tensor_parallel_size=4,
+        rbln_config={
+            "vlm": {
+                "language_model": {
+                    # For the definition of prefill_chunk_size, please refer to the link below.
+                    # https://docs.rbln.ai/v0.9.4/software/optimum/model_api/transformers/text_model/common.html#optimum.rbln.transformers.models.decoderonly.configuration_decoderonly.RBLNDecoderOnlyModelForCausalLMConfig.__init__
+                    # default "max_position_embeddings"
+                    "prefill_chunk_size": 8192,
+                },
+            }
+        },
     )
 
     # Save compiled results to disk